[Refactor] #134 - ServiceProtocol을 Repository로 수정 후 레이어 변경
