+++ CERTI-iOS/Domain/Interfaces/Repositories/CareersRepository.swift
@@ -1,0 +1,16 @@
+//
+//  CareersRepository.swift
+//  CERTI-iOS
+//
+//  Created by OneTen on 8/15/25.
+//
+
+
+import Foundation
+import Moya
+
+protocol CareersRepository {
+    func fetchCareersList() async -> Result<CareersListResponseDTO, NetworkError>
+    func deledteCareers(id: Int) async -> Result<Void, NetworkError>
+    func addCareer(request: AddCareerRequestDTO) async -> Result<Bool, NetworkError>
+}
